+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [UIMARCAUTH-178](https://issues.folio.org/browse/UIMARCAUTH-178) MARC authority: Delete MARC authority record handling
 - [UIMARCAUTH-222](https://issues.folio.org/browse/UIMARCAUTH-222) MARC Authority: Edit MARC authority record
 - [UIMARCAUTH-223](https://issues.folio.org/browse/UIMARCAUTH-223) bump stripes to 8.0.0 for Orchid/2023-R1
+- [UIMARCAUTH-220](https://issues.folio.org/browse/UIMARCAUTH-220) Any user assign the permission View Authority record permissions can export authority reports
 
 ## [2.0.1](https://github.com/folio-org/ui-marc-authorities/tree/v2.0.1) (2022-11-28)